autorizacao para criar historico

+++ shopList/src/models/ItemCupomModel.ts
@@ -39,7 +39,7 @@ export default (sequelize: Sequelize.Sequelize, DataTypes: Sequelize.DataTypes) 
                 allowNull: false,                
             },
             qtde:{
-                type: DataTypes.INTEGER,
+                type: DataTypes.DECIMAL(10,4),
                 allowNull: true
             },
             unidade:{
